Outsources moment computation (area and centroid) in BezierCurve and CurvedPolygon
These are now provided as primal operators:
* BezierCurve: `sector_area()` / `sector_centroid()`
* CurvedPolygon: `area()` / `centroid()`

There operations only apply to 2D (polynomial) BezierCurves, so don't belong
in the dimension-templated classes.
