+++ src/axom/primal/tests/primal_bezier_curve.cpp
@@ -204,88 +204,6 @@ TEST(primal_beziercurve_, tangent)
 }
 
 //------------------------------------------------------------------------------
-TEST(primal_beziercurve, sector_area_cubic)
-{
-  const int DIM = 2;
-  using CoordType = double;
-  using PointType = primal::Point<CoordType, DIM>;
-  using BezierCurveType = primal::BezierCurve<CoordType, DIM>;
-
-  {
-    SLIC_INFO("Testing Bezier sector area calculation for a cubic");
-    const int order = 3;
-    PointType data[order + 1] = {PointType {0.6, 1.2},
-                                 PointType {1.3, 1.6},
-                                 PointType {2.9, 2.4},
-                                 PointType {3.2, 3.5}};
-
-    BezierCurveType bCurve(data, order);
-    EXPECT_TRUE(axom::utilities::isNearlyEqual(bCurve.sectorArea(), .1455));
-  }
-}
-
-//------------------------------------------------------------------------------
-TEST(primal_beziercurve, sector_moment_cubic)
-{
-  const int DIM = 2;
-  using CoordType = double;
-  using PointType = primal::Point<CoordType, DIM>;
-  using BezierCurveType = primal::BezierCurve<CoordType, DIM>;
-
-  {
-    SLIC_INFO("Testing Bezier sector moment calculation for a cubic");
-    const int order = 3;
-    PointType data[order + 1] = {PointType {0.6, 1.2},
-                                 PointType {1.3, 1.6},
-                                 PointType {2.9, 2.4},
-                                 PointType {3.2, 3.5}};
-
-    BezierCurveType bCurve(data, order);
-    PointType M = bCurve.sectorCentroid();
-    EXPECT_NEAR(M[0], -.429321428571429, 2e-15);
-    EXPECT_NEAR(M[1], -.354010714285715, 2e-15);
-  }
-}
-
-//------------------------------------------------------------------------------
-TEST(primal_beziercurve, sector_area_point)
-{
-  const int DIM = 2;
-  using CoordType = double;
-  using PointType = primal::Point<CoordType, DIM>;
-  using BezierCurveType = primal::BezierCurve<CoordType, DIM>;
-
-  {
-    SLIC_INFO("Testing Bezier sector area calculation for a point");
-    const int order = 0;
-    PointType data[order + 1] = {PointType::make_point(0.6, 1.2)};
-
-    BezierCurveType bCurve(data, order);
-    EXPECT_DOUBLE_EQ(bCurve.sectorArea(), 0.0);
-  }
-}
-
-//------------------------------------------------------------------------------
-TEST(primal_beziercurve, sector_moment_point)
-{
-  const int DIM = 2;
-  using CoordType = double;
-  using PointType = primal::Point<CoordType, DIM>;
-  using BezierCurveType = primal::BezierCurve<CoordType, DIM>;
-
-  {
-    SLIC_INFO("Testing Bezier sector moment calculation for a point");
-    const int order = 0;
-    PointType data[order + 1] = {PointType::make_point(0.6, 1.2)};
-
-    BezierCurveType bCurve(data, order);
-    PointType M = bCurve.sectorCentroid();
-    EXPECT_DOUBLE_EQ(M[0], 0.0);
-    EXPECT_DOUBLE_EQ(M[1], 0.0);
-  }
-}
-
-//------------------------------------------------------------------------------
 TEST(primal_beziercurve, split_cubic)
 {
   SLIC_INFO("Testing Bezier splitting of a cubic");
@@ -521,133 +439,6 @@ TEST(primal_beziercurve, isLinear)
       const double tol = 0.01;
       const double tol_sq = tol * tol;
       EXPECT_FALSE(curve.isLinear(tol_sq));
-    }
-  }
-}
-
-TEST(primal_beziercurve, sector_weights)
-{
-  SLIC_INFO("Testing weights for BezierCurve::sectorArea()");
-
-  // NOTE: Expected weights are provided in the reference paper [Ueda99]
-  // See doxygen comment for BezierCurve::sectorArea()
-
-  using CoordType = double;
-
-  primal::internal::MemoizedSectorAreaWeights<CoordType> memoizedSectorWeights;
-
-  // order 1
-  {
-    const int ord = 1;
-    auto weights = memoizedSectorWeights.getWeights(ord);
-
-    double binomInv = 1. / axom::utilities::binomialCoefficient(2, 1);
-    axom::numerics::Matrix<CoordType> exp(ord + 1, ord + 1);
-    // clang-format off
-    exp(0,0) =  0; exp(0,1) =  1;
-    exp(1,0) = -1; exp(1,1) =  0;
-    // clang-format on
-
-    for(int i = 0; i <= ord; ++i)
-    {
-      for(int j = 0; j <= ord; ++j)
-      {
-        EXPECT_DOUBLE_EQ(exp(i, j) * binomInv, weights(i, j));
-      }
-    }
-  }
-
-  // order 2
-  {
-    const int ord = 2;
-    auto weights = memoizedSectorWeights.getWeights(ord);
-
-    double binomInv = 1. / axom::utilities::binomialCoefficient(4, 2);
-    axom::numerics::Matrix<CoordType> exp(ord + 1, ord + 1);
-    // clang-format off
-    exp(0,0) =  0; exp(0,1) =  2; exp(0,2) =  1;
-    exp(1,0) = -2; exp(1,1) =  0; exp(1,2) =  2;
-    exp(2,0) = -1; exp(2,1) = -2; exp(2,2) =  0;
-    // clang-format on
-
-    for(int i = 0; i <= ord; ++i)
-    {
-      for(int j = 0; j <= ord; ++j)
-      {
-        EXPECT_DOUBLE_EQ(exp(i, j) * binomInv, weights(i, j));
-      }
-    }
-  }
-
-  // order 3
-  {
-    const int ord = 3;
-    auto weights = memoizedSectorWeights.getWeights(ord);
-
-    double binomInv = 1. / axom::utilities::binomialCoefficient(6, 3);
-    axom::numerics::Matrix<CoordType> exp(ord + 1, ord + 1);
-    // clang-format off
-    exp(0,0) =  0; exp(0,1) =  6; exp(0,2) =  3; exp(0,3) =  1;
-    exp(1,0) = -6; exp(1,1) =  0; exp(1,2) =  3; exp(1,3) =  3;
-    exp(2,0) = -3; exp(2,1) = -3; exp(2,2) =  0; exp(2,3) =  6;
-    exp(3,0) = -1; exp(3,1) = -3; exp(3,2) = -6; exp(3,3) =  0;
-    // clang-format on
-
-    for(int i = 0; i <= ord; ++i)
-    {
-      for(int j = 0; j <= ord; ++j)
-      {
-        EXPECT_DOUBLE_EQ(exp(i, j) * binomInv, weights(i, j));
-      }
-    }
-  }
-
-  // order 4
-  {
-    const int ord = 4;
-    auto weights = memoizedSectorWeights.getWeights(ord);
-
-    double binomInv = 1. / axom::utilities::binomialCoefficient(8, 4);
-    axom::numerics::Matrix<CoordType> exp(ord + 1, ord + 1);
-    // clang-format off
-    exp(0,0) =  0; exp(0,1) = 20; exp(0,2) = 10; exp(0,3) =  4; exp(0,4) =  1;
-    exp(1,0) =-20; exp(1,1) =  0; exp(1,2) =  8; exp(1,3) =  8; exp(1,4) =  4;
-    exp(2,0) =-10; exp(2,1) = -8; exp(2,2) =  0; exp(2,3) =  8; exp(2,4) = 10;
-    exp(3,0) = -4; exp(3,1) = -8; exp(3,2) = -8; exp(3,3) =  0; exp(3,4) = 20;
-    exp(4,0) = -1; exp(4,1) = -4; exp(4,2) =-10; exp(4,3) =-20; exp(4,4) =  0;
-    // clang-format on
-
-    for(int i = 0; i <= ord; ++i)
-    {
-      for(int j = 0; j <= ord; ++j)
-      {
-        EXPECT_DOUBLE_EQ(exp(i, j) * binomInv, weights(i, j));
-      }
-    }
-  }
-
-  // order 5
-  {
-    const int ord = 5;
-    auto weights = memoizedSectorWeights.getWeights(ord);
-
-    double binomInv = 1. / axom::utilities::binomialCoefficient(10, 5);
-    axom::numerics::Matrix<CoordType> exp(ord + 1, ord + 1);
-    // clang-format off
-    exp(0,0) =  0; exp(0,1) = 70; exp(0,2) = 35; exp(0,3) = 15; exp(0,4) =  5; exp(0,5) =  1;
-    exp(1,0) =-70; exp(1,1) =  0; exp(1,2) = 25; exp(1,3) = 25; exp(1,4) = 15; exp(1,5) =  5;
-    exp(2,0) =-35; exp(2,1) =-25; exp(2,2) =  0; exp(2,3) = 20; exp(2,4) = 25; exp(2,5) = 15;
-    exp(3,0) =-15; exp(3,1) =-25; exp(3,2) =-20; exp(3,3) =  0; exp(3,4) = 25; exp(3,5) = 35;
-    exp(4,0) = -5; exp(4,1) =-15; exp(4,2) =-25; exp(4,3) =-25; exp(4,4) =  0; exp(4,5) = 70;
-    exp(5,0) = -1; exp(5,1) = -5; exp(5,2) =-15; exp(5,3) =-35; exp(5,4) =-70; exp(5,5) =  0;
-    // clang-format on
-
-    for(int i = 0; i <= ord; ++i)
-    {
-      for(int j = 0; j <= ord; ++j)
-      {
-        EXPECT_DOUBLE_EQ(exp(i, j) * binomInv, weights(i, j));
-      }
     }
   }
 }
